% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{betas.slice.sampling}
\alias{betas.slice.sampling}
\title{betas.slice.sampling}
\usage{
betas.slice.sampling(Sigma.b, Y, X, betas, alpha, C, lambda, w, m, form)
}
\arguments{
\item{Sigma.b}{variance estimate of betas}

\item{Y}{response variable}

\item{X}{covariates for betas}

\item{betas}{current value of betas}

\item{alpha}{probability of true censoring}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for betas
}
