% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{lambda.slice.sampling}
\alias{lambda.slice.sampling}
\title{lambda.slice.sampling}
\usage{
lambda.slice.sampling(Y, eXB, alpha, C, lambda, w, m, lower = 0.01,
  upper = +Inf)
}
\arguments{
\item{Y}{response variable}

\item{eXB}{exponentiated vector of covariates times betas}

\item{alpha}{probability of true censoring}

\item{C}{censoring indicator}

\item{lambda}{current value of lambda}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for lambda
}
