% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{univ.V.slice.sampling}
\alias{univ.V.slice.sampling}
\title{univ.V.slice.sampling}
\usage{
univ.V.slice.sampling(V.s, s, S, A, lambda, Y, eXB, Z, V, gammas, C, rho, w, m,
  lower = -Inf, upper = +Inf, form)
}
\arguments{
\item{V.s}{current value of the sth element of V}

\item{s}{sth element}

\item{S}{spatial information (e.g. district)}

\item{A}{adjacency information corresponding to spatial information}

\item{lambda}{CAR parameter}

\item{Y}{response variable}

\item{eXB}{exponentiated vector of covariates times betas}

\item{Z}{covariates for gammas}

\item{V}{spatial random effects}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for gammas.p
}
