% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{univ.W.slice.sampling}
\alias{univ.W.slice.sampling}
\title{univ.W.slice.sampling}
\usage{
univ.W.slice.sampling(W.s, s, S, A, lambda, Y, X, W, betas, delta, C, rho, w, m,
  lower = -Inf, upper = +Inf, form)
}
\arguments{
\item{W.s}{current value of the sth element of W}

\item{s}{sth element}

\item{S}{spatial information (e.g. district)}

\item{A}{adjacency information corresponding to spatial information}

\item{lambda}{CAR parameter}

\item{Y}{response variable}

\item{X}{covariates for betas}

\item{W}{spatial random effects}

\item{betas}{current value of betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for W.s
}
