% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{V.post}
\alias{V.post}
\title{V.post}
\usage{
V.post(V.s, s, S, A, lambda, Y, eXB, Z, V, gammas, C, rho, form)
}
\arguments{
\item{V.s}{current value of the sth element of V}

\item{s}{sth element}

\item{S}{spatial information (e.g. district)}

\item{A}{adjacency information corresponding to spatial information}

\item{lambda}{CAR parameter}

\item{Y}{response variable}

\item{eXB}{exponentiated vector of covariates times betas}

\item{Z}{covariates for gammas}

\item{V}{spatial random effects}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of V with sth element fixed as V.s
}
