% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{V.slice.sampling}
\alias{V.slice.sampling}
\title{V.slice.sampling}
\usage{
V.slice.sampling(S, A, lambda, Y, eXB, Z, V, gammas, C, rho, w, m,
  form = form)
}
\arguments{
\item{S}{spatial information (e.g. district)}

\item{A}{adjacency information corresponding to spatial information}

\item{lambda}{CAR parameter}

\item{Y}{response variable}

\item{eXB}{exponentiated vector of covariates times betas}

\item{Z}{covariates for gammas}

\item{V}{spatial random effects}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for W
}
