% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{W.slice.sampling}
\alias{W.slice.sampling}
\title{W.slice.sampling}
\usage{
W.slice.sampling(S, A, lambda, Y, X, W, betas, delta, C, rho, w, m,
  form = form)
}
\arguments{
\item{S}{spatial information (e.g. district)}

\item{A}{adjacency information corresponding to spatial information}

\item{lambda}{CAR parameter}

\item{Y}{response variable}

\item{X}{covariates for betas}

\item{W}{spatial random effects}

\item{betas}{current value of betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{w}{size of the slice in the slice sampling}

\item{m}{limit on steps in the slice sampling}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for W
}
